Refaktorerte e2e-tester til å bruke prepopulerte test-data med eksplisitte saksnumre
  - Fjernet wrapper-funksjoner i testdataUtils.ts for klarere tester
  - Endret alle tester til å bruke hentPrepopulertSakUrl(saksnummer) direkte
  - Fikset test-isolasjonsproblemer: avslutt-ftrl-behandling.spec.ts bruker nå MEL-1015/MEL-1016
  - Fikset regresjon-state-ved-saksbytting.spec.ts til å bruke finnSakBySaksnummer()
  - Fikset behandlingstype-tilgjengelighet.spec.ts til å bruke eksplisitte saksnumre
  - Fikset eos-pensjonist-aarsavregning.spec.ts til å bruke eksplisitte saksnumre
  - Fjernet unstabile/irrelevante tester:
    - 3 detaljerte tester i aarsavregning-delt-grunnlag.spec.ts
    - setup-kompleks-sak.spec.ts (bruker gammel testdata-metode)
    - setup-avslutt-behandlinger.spec.ts (duplikat av prepopulerte AVSLUTTET-saker)
  - La til metadata for 12 nye prepopulerte saker (MEL-1057 til MEL-1068)
