+++ tests/e2e/specs/opprett-ny-sak/ny-sak/bruker-sak-validering.spec.ts
@@ -1,4 +1,4 @@
-import { test, Page } from "@playwright/test";
+import { Page, test } from "@playwright/test";
 import { runAxeAnalyze } from "../../../utils/axeUtils";
 import { HovedsidePage, USER_ID_VALID } from "../../../pages/hovedside.page";
 import { OpprettNySakPage } from "../../../pages/opprett-ny-sak/opprett-ny-sak.page";
@@ -17,7 +17,11 @@ async function setupOpprettNySakTester(page: Page) {
 
 test.describe("'Opprett ny sak for bruker", () => {
   test.beforeEach(async ({ page }) => {
-    await setupOpprettNySakTester(page);
+    const mainPage = new HovedsidePage(page);
+    opprettNySakPage = new OpprettNySakPage(page);
+
+    await mainPage.goto();
+    await mainPage.klikkOpprettNySakKnapp();
   });
 
   test("Manglende påkrevde felt - viser feilmeldinger", async ({ page }, testInfo) => {
